feat: Explicitly reject unsupported shapes, traits, etc. (#257)
*Issue #, if available:* 

#185

*Description of changes:*

Adds a model validator with allow-lists for shapes and traits the code generators understand, meaning that we are aware of the meaning of the shape or trait and believe that the correct code generation is implemented. For many service traits this just means we've confirmed that they don't change the API callers use when in `--aws-sdk` style mode, but some such as `@streaming` definitely do affect this API and hence are not in the list.

We could have specific lists for individual target languages as well in the future, but this will be more difficult to fit into the validation framework, which isn't aware of which target language we're building for. Dividing up the codegen into separate plugins with their own validators would be the natural way to address this.

I populated these lists by adding all necessary entries to cover the TestModels, but also tested on all downstream projects via `make polymorph_dafny`, which did reveal a couple of tweaks not covered by the test models.

Resolves #185