fix(rust): validate config on client construction (#748)
Description of changes:
In the Rust codegen, constraint validation is performed on operation calls instead of within structure builders' build() function. (The reasoning for this choice is documented in the description of #582.) But the validation wasn't applied to the constructors of localService clients, so attempting to construct a client with an invalid config could panic during conversion (in particular, when a @required field was missing).

This PR implements the missing validation in client constructors, and tests that both valid and invalid configs have the expected behavior when passed to the client constructor.